refactor: ratchet app physical gates to 260 and densify cap tests (#4582)
Wave-20 Slice E enforces the 260-line lib ceiling after host splits, pins
checker ACs, and restores headroom on the four 400-line app tests.
