+++ SPEC/program/app-core-services-file-size.md
@@ -23,8 +23,8 @@
 
 - **Physical lines:** `LineSplitter` line count of the file’s UTF-8 text (same
   as `const LineSplitter().convert(content).length` in the checker).
-- **Threshold:** each scanned file must have **300** physical lines or fewer
-  (301 or more fails).
+- **Threshold:** each scanned file must have **260** physical lines or fewer
+  (261 or more fails). Wave-20 #4582 lowered the wave-16 **300** cap.
 
 There is **no** YAML, keyed table, or per-file exemption that raises the
 effective cap for a specific path. The grandfather list stays empty.
@@ -32,14 +32,14 @@ effective cap for a specific path. The grandfather list stays empty.
 ## Acceptance criteria
 
 - Given a temporary workspace that contains
-  `app/lib/core/services/game_service/too_long.dart` with **301** physical lines
+  `app/lib/core/services/game_service/too_long.dart` with **261** physical lines
   and no other violating files, when the System runs
   `runCheckAppCoreServicesFileSize` with that workspace root, then the checker
   exits non-zero and the error output names `too_long.dart` and reports
-  `301 physical lines > 300`.
+  `261 physical lines > 260`.
 
 - Given a temporary workspace whose only matching file is
-  `app/lib/core/services/game_service/ok.dart` with exactly **300** physical
+  `app/lib/core/services/game_service/ok.dart` with exactly **260** physical
   lines, when the System runs `runCheckAppCoreServicesFileSize`, then the
   checker exits zero.
