+++ test/check_app_catalog_widgets_file_size_test.dart
@@ -5,7 +5,11 @@ import 'package:test/test.dart';
 import '../tool/check_app_catalog_widgets_file_size.dart';
 
 void main() {
-  test('fails when a catalog widget file exceeds 300 lines', () {
+  test('wave-20 ceiling is 260', () {
+    expect(appCatalogWidgetsFileSizeCeiling, 260);
+  });
+
+  test('fails when a catalog widget file exceeds 260 lines', () {
     final temp = Directory.systemTemp.createTempSync(
       'check_app_catalog_widgets_file_size_fail_',
     );
@@ -13,7 +17,7 @@ void main() {
 
     final violatingFile = File('${temp.path}/app/lib/widgets/huge_ct.dart')
       ..createSync(recursive: true);
-    violatingFile.writeAsStringSync(List.filled(301, '// line').join('\n'));
+    violatingFile.writeAsStringSync(List.filled(261, '// line').join('\n'));
 
     final logs = <String>[];
     final code = runCheckAppCatalogWidgetsFileSize(
@@ -24,7 +28,7 @@ void main() {
 
     expect(code, 1);
     expect(logs.join('\n'), contains('huge_ct.dart'));
-    expect(logs.join('\n'), contains('301 physical lines > 300'));
+    expect(logs.join('\n'), contains('261 physical lines > 260'));
   });
 
   test('fails when catalog widgets directory is missing', () {
@@ -44,7 +48,7 @@ void main() {
     expect(logs.join('\n'), contains('widgets not found'));
   });
 
-  test('passes when all catalog widget files are at or below 300 lines', () {
+  test('passes when all catalog widget files are at or below 260 lines', () {
     final temp = Directory.systemTemp.createTempSync(
       'check_app_catalog_widgets_file_size_pass_',
     );
@@ -52,7 +56,7 @@ void main() {
 
     final okFile = File('${temp.path}/app/lib/widgets/panel.dart')
       ..createSync(recursive: true);
-    okFile.writeAsStringSync(List.filled(300, '// line').join('\n'));
+    okFile.writeAsStringSync(List.filled(260, '// line').join('\n'));
 
     final code = runCheckAppCatalogWidgetsFileSize(temp.path);
     expect(code, 0);
