+++ test/check_app_features_screens_lib_physical_file_size_test.dart
@@ -7,6 +7,11 @@ import '../tool/check_app_features_screens_lib_physical_file_size.dart';
 const _screensRel = 'app/lib/features/game/screens';
 
 void main() {
+  test('wave-20 ceiling is 260', () {
+    expect(appFeaturesScreensLibPhysicalFileSizeCeiling, 260);
+    expect(appFeaturesScreensLibPhysicalFileSizeGrandfatheredForTests, isEmpty);
+  });
+
   test('passes for the real app feature screens subtree', () {
     final logs = <String>[];
     final code = runCheckAppFeaturesScreensLibPhysicalFileSize(
@@ -34,7 +39,7 @@ void main() {
     Directory('${temp.path}/$_screensRel').createSync(recursive: true);
     File('${temp.path}/$_screensRel/huge.dart')
       ..createSync(recursive: true)
-      ..writeAsStringSync(List.filled(301, 'final x = 1;').join('\n'));
+      ..writeAsStringSync(List.filled(261, 'final x = 1;').join('\n'));
 
     final logs = <String>[];
     final code = runCheckAppFeaturesScreensLibPhysicalFileSize(
@@ -46,7 +51,7 @@ void main() {
 
     expect(code, 1);
     expect(logs.join('\n'), contains('huge.dart'));
-    expect(logs.join('\n'), contains('physical lines > 300'));
+    expect(logs.join('\n'), contains('physical lines > 260'));
   });
 
   test('fails when the screens directory is missing', () {
