+++ test/check_app_turn_resolution_file_size_test.dart
@@ -5,7 +5,11 @@ import 'package:test/test.dart';
 import '../tool/check_app_turn_resolution_file_size.dart';
 
 void main() {
-  test('fails when a turn_resolution file exceeds 300 lines', () {
+  test('wave-20 ceiling is 260', () {
+    expect(appTurnResolutionFileSizeCeiling, 260);
+  });
+
+  test('fails when a turn_resolution file exceeds 260 lines', () {
     final temp = Directory.systemTemp.createTempSync(
       'check_app_turn_resolution_file_size_fail_',
     );
@@ -14,7 +18,7 @@ void main() {
     final violatingFile = File(
       '${temp.path}/app/lib/features/game/turn_resolution/too_long.dart',
     )..createSync(recursive: true);
-    violatingFile.writeAsStringSync(List.filled(301, '// line').join('\n'));
+    violatingFile.writeAsStringSync(List.filled(261, '// line').join('\n'));
 
     final logs = <String>[];
     final code = runCheckAppTurnResolutionFileSize(
@@ -25,7 +29,7 @@ void main() {
 
     expect(code, 1);
     expect(logs.join('\n'), contains('too_long.dart'));
-    expect(logs.join('\n'), contains('301 physical lines > 300'));
+    expect(logs.join('\n'), contains('261 physical lines > 260'));
   });
 
   test('fails when turn_resolution directory is missing', () {
@@ -45,7 +49,7 @@ void main() {
     expect(logs.join('\n'), contains('turn_resolution not found'));
   });
 
-  test('passes when all turn_resolution files are at or below 300 lines', () {
+  test('passes when all turn_resolution files are at or below 260 lines', () {
     final temp = Directory.systemTemp.createTempSync(
       'check_app_turn_resolution_file_size_pass_',
     );
@@ -54,7 +58,7 @@ void main() {
     final okFile = File(
       '${temp.path}/app/lib/features/game/turn_resolution/ok.dart',
     )..createSync(recursive: true);
-    okFile.writeAsStringSync(List.filled(300, '// line').join('\n'));
+    okFile.writeAsStringSync(List.filled(260, '// line').join('\n'));
 
     final code = runCheckAppTurnResolutionFileSize(temp.path);
     expect(code, 0);
