+++ tool/check_game_widgets_file_size.dart
@@ -5,11 +5,11 @@ import 'dart:io';
 
 import 'package:path/path.dart' as p;
 
-/// Ratchet ceiling for game feature widgets (wave-15 #4352 Slice B).
-const int gameWidgetsFileSizeCeiling = 300;
+/// Ratchet ceiling for game feature widgets (wave-20 #4582; was 300).
+const int gameWidgetsFileSizeCeiling = 260;
 
 /// PR-blocking structural check: files under
-/// `app/lib/features/game/widgets/**` must stay at or below 300 physical lines.
+/// `app/lib/features/game/widgets/**` must stay at or below 260 physical lines.
 int runCheckGameWidgetsFileSize(
   String repoRoot, {
   void Function(String line)? info,
