+++ tool/ct_repo_lint_manifest.yaml
@@ -697,7 +697,7 @@ rules:
 
   - rule_id: repo.app_flame_lib_physical_file_size
     group: structure
-    title: app/lib/features/game/flame files must stay at or below 300 physical lines (shrink-only grandfather)
+    title: app/lib/features/game/flame files must stay at or below 260 physical lines (wave-20; empty grandfather; Refs #4582)
     spec: SPEC/program/repo-lint.md
     runner: dart
     script: tool/check_app_flame_lib_physical_file_size.dart
@@ -705,7 +705,7 @@ rules:
 
   - rule_id: repo.app_features_screens_lib_physical_file_size
     group: structure
-    title: app/lib/features/game/screens files must stay at or below 300 physical lines (shrink-only grandfather)
+    title: app/lib/features/game/screens files must stay at or below 260 physical lines (wave-20; empty grandfather; Refs #4582)
     spec: SPEC/program/app-features-screens-file-size.md
     runner: dart
     script: tool/check_app_features_screens_lib_physical_file_size.dart
@@ -1205,14 +1205,14 @@ rules:
 
   - rule_id: repo.game_widgets_file_size
     group: structure
-    title: app/lib/features/game/widgets files must stay at or below 300 physical lines
+    title: app/lib/features/game/widgets files must stay at or below 260 physical lines (Refs #4582)
     spec: SPEC/program/game-widgets-file-size.md
     runner: dart
     script: tool/check_game_widgets_file_size.dart
 
   - rule_id: repo.app_catalog_widgets_file_size
     group: structure
-    title: app/lib/widgets catalog files must stay at or below 300 physical lines (Refs #4352)
+    title: app/lib/widgets catalog files must stay at or below 260 physical lines (Refs #4352, #4582)
     spec: SPEC/program/app-catalog-widgets-file-size.md
     runner: dart
     script: tool/check_app_catalog_widgets_file_size.dart
@@ -1220,7 +1220,7 @@ rules:
 
   - rule_id: repo.app_core_services_file_size
     group: structure
-    title: app/lib/core/services files must stay at or below 300 physical lines (Refs #4450)
+    title: app/lib/core/services files must stay at or below 260 physical lines (Refs #4450, #4582)
     spec: SPEC/program/app-core-services-file-size.md
     runner: dart
     script: tool/check_app_core_services_file_size.dart
@@ -1228,7 +1228,7 @@ rules:
 
   - rule_id: repo.app_turn_resolution_file_size
     group: structure
-    title: app/lib/features/game/turn_resolution files must stay at or below 300 physical lines (Refs #4512)
+    title: app/lib/features/game/turn_resolution files must stay at or below 260 physical lines (Refs #4512, #4582)
     spec: SPEC/program/app-turn-resolution-file-size.md
     runner: dart
     script: tool/check_app_turn_resolution_file_size.dart
